# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lorekeeper Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List Campaigns",
        "operationId": "list_campaigns",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponse"
                  },
                  "type": "array",
                  "title": "Response List Campaigns"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{slug}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign",
        "operationId": "get_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Patch Campaign",
        "operationId": "patch_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete Campaign",
        "operationId": "delete_campaign",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{slug}/characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "List Characters",
        "operationId": "list_characters",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "character_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CharacterType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Character Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponse"
                  },
                  "title": "Response List Characters"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "summary": "Create Character",
        "operationId": "create_character",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/v1/campaigns/{slug}/characters/{character_slug}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "Get Character",
        "operationId": "get_character",
        "parameters": [
          {
            "name": "character_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Characters"
        ],
        "summary": "Patch Character",
        "operationId": "patch_character",
        "parameters": [
          {
            "name": "character_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCharacterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "summary": "Delete Character",
        "operationId": "delete_character",
        "parameters": [
          {
            "name": "character_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{slug}/characters/{character_slug}/image": {
      "put": {
        "tags": [
          "Characters"
        ],
        "summary": "Upload Character Image",
        "operationId": "upload_character_image",
        "parameters": [
          {
            "name": "character_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_character_image"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "Image is not a supported type or exceeds the size limit"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "summary": "Delete Character Image",
        "operationId": "delete_character_image",
        "parameters": [
          {
            "name": "character_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Slug"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "invite_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "slug",
          "role",
          "invite_code",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse",
        "example": {
          "created_at": "2024-01-15T10:00:00Z",
          "description": "A gothic horror campaign set in the mist-shrouded land of Barovia.",
          "id": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
          "invite_code": "dX9kLmN2pQrS4tUvWxYz",
          "name": "Curse of Strahd",
          "role": "gm",
          "slug": "curse-of-strahd-a1b2c3d4",
          "updated_at": "2024-01-15T10:00:00Z"
        }
      },
...
      "CharacterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "character_type": {
            "$ref": "#/components/schemas/CharacterType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "character_type",
          "description",
          "restricted",
          "tags",
          "image_url",
          "created_at",
          "updated_at"
        ],
        "title": "CharacterResponse",
        "example": {
          "character_type": "pc",
          "created_at": "2024-01-15T10:00:00Z",
          "description": "A wise elven druid from the Emerald Enclave.",
          "id": "a4c2b1d0-e3f5-4789-abcd-ef0123456789",
          "image_url": "/media/3f9c1e2a-3b7e-4a2e-9b1a-9d6a2b0e5c11.jpg",
          "name": "Elara Moonwhisper",
          "restricted": false,
          "slug": "elara-moonwhisper",
          "tags": [
            "ally",
            "spellcaster"
          ],
          "updated_at": "2024-01-15T10:00:00Z"
        }
      },
      "CharacterType": {
        "type": "string",
        "enum": [
          "pc",
          "npc"
        ],
        "title": "CharacterType"
      },
...
      "CreateCampaignRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug_label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug Label"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug_label"
        ],
        "title": "CreateCampaignRequest",
        "example": {
          "description": "A gothic horror campaign set in the mist-shrouded land of Barovia.",
          "name": "Curse of Strahd",
          "slug_label": "curse-of-strahd"
        }
      },
      "CreateCharacterRequest": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "character_type": {
            "$ref": "#/components/schemas/CharacterType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "character_type"
        ],
        "title": "CreateCharacterRequest",
        "example": {
          "character_type": "pc",
          "description": "A wise elven druid from the Emerald Enclave.",
          "name": "Elara Moonwhisper",
          "slug": "elara-moonwhisper",
          "tags": [
            "ally",
            "spellcaster"
          ]
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchCampaignRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slug_label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*\\z",
            "title": "Slug Label"
          }
        },
        "type": "object",
        "title": "PatchCampaignRequest",
        "example": {
          "description": "Updated campaign description.",
          "name": "Curse of Strahd: Reloaded",
          "slug_label": "curse-of-strahd-reloaded"
        }
      },
      "PatchCharacterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "character_type": {
            "$ref": "#/components/schemas/CharacterType",
            "title": "Character Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "restricted": {
            "type": "boolean",
            "title": "Restricted"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "PatchCharacterRequest",
        "example": {
          "character_type": "pc",
          "description": "Updated description.",
          "name": "Elara of the Enclave",
          "tags": [
            "ally",
            "spellcaster"
          ]
        }
      },
...
    }
  },
...
}